App: fix external document loading
The problem happens when partial loading is enabled. If document A
contains a link to some object in document B, it will load B as partial
document with only that object and its necessary dependencies. But if
document A contains another link to some object in document C which
also has a link to some object in document B, the link in document C
may not be restored, because document B is partially loaded without
the linked object. This patch will check for this case and reload
document B for more objects.

See an example reported in
https://forum.freecadweb.org/viewtopic.php?p=495078#p495078