App/Gui: change Link copy on change behavior
Use enum to represent different copy on change mode.

Change 'Owned' mode to force copy of the configurable object.

Change the way the copied objects are kept, by using a 'hidden'
LinkGroup, so as to make sure none of the copied object are exposed to
the root of the object hierarchy.