Gui: fix action label in status bar
Make it stretch as before, so that the status message can get the proper
size. Note QStatusBar shows the message (warning, error message in
FreeCAD) in its own widget, while we display pre-selection message in a
QLabel inside the QStatusBar.  However, we use the QLabel width to
truncate the the status bar message with QFontMetrics::elidedText().